pyproject.toml
[project]
name = "ruian-import"
version = "0.1.0"
description = "RUIAN data import to PostGIS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "lxml>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
pdf = [
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",
]

[project.scripts]
ruian-download = "scripts.download_ruian:main"
ruian-import = "scripts.import_ruian:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ruian_import", "src/notice_boards", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/ruian_import --cov=src/notice_boards --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["psycopg2", "psycopg2.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["fitz", "pdfplumber"]
ignore_missing_imports = true
